User Story E2 & E4 (#16)
* MODIFIED class diagram

* ADDED cashier view + functionality

* FIXED bugs

* ADDED error message "wrong barcode"

* ADDED javadoc

* ADDED code cleanup

* MERGED rebased 'beniamin' onto 'main'

* User Stories E1, E2, E4 backend bugfix

- ADDED incomplete unit testing for scanItem method
- FIXED message for thrown exception in scanItem method

* User Stories E1, E2, E4 backend bugfix

- MODIFIED now the view model receives items with quantity 1 in the arraylist

* FIXED item quantity

* User Stories D2 R1 backend + E1, E2, E4 frontend bugfix

- FIXED major bug with items, now price is of type double and id is of type long
- FIXED now cashier view shows pop-up errors given an exception
- FIXED typo and small code irregularities

+++ Server/src/mediator/ServerNetworkManager.java
@@ -119,7 +119,7 @@ public class ServerNetworkManager implements RemoteModel {
     }
 
     @Override
-    public void changePrice(int id, int price) {
+    public void changePrice(Long id, Double price) {
         serverModel.changePrice(id, price);
     }
