+++ Server/src/model/ServerInventoryModel.java
@@ -4,5 +4,6 @@ import java.util.ArrayList;
 
 public interface ServerInventoryModel {
     ArrayList<Item> getItems();
-    void changePrice(int id, int price);
+
+    void changePrice(Long id, Double price);
 }
