+++ Shared/src/networking/RemoteInventoryModel.java
@@ -8,5 +8,6 @@ import java.util.ArrayList;
 
 public interface RemoteInventoryModel extends Remote {
     ArrayList<Item> getItems() throws RemoteException;
-    void changePrice(int id,int price) throws RemoteException;
+
+    void changePrice(Long id, Double price) throws RemoteException;
 }
